dependency: Add runtime cache
Dependencies found in a subproject fallback should always return the
same dependency object in subsequent calls even if fallback kwarg is
omitted. Likewise a not-found dependency should stay not-found even if
subsequent calls provide a fallback.

This avoids having parts of a project linking on a system library and
other parts linking on a subproject library.