WIP: PkgConfigDependency: Use internal pkg-config implementation
Unless PKG_CONFIG is set in env or pkgconfig is set in machine file, in
which case we have to use the binary provided by user because it could
be a wrapper that sets some env.

Add internal `meson pkg-config` command line mimicking pkg-config CLI.
It is far from complete and only expose features required by
PkgConfigDependency.