wrap: Add support do download into a destination directory
It is common for projects to want to run their CI in offline mode, or at
least reduce as much as possible download time during each job. Some
projects uses `meson subprojects download` during docker image
generation and store the `subprojects/` directory directly into the
docker image. When a job is run, the cache is copied back into
subprojects/ directory.

One difficulty with that workflow is it is difficult to exclude
subprojects that are part of the main project's git repository. Keeping
only tarballs for wrap-file also helps reducing docker image size.