compilers: Differentiate .S and .s
The case is important: .S files needs to be preprocessed by a C-like
compiler and .s can be compiled directly using GAS-like assembler.

Currently assume that all compilers that can do .S can also do .s. In
the future MSVCCompiler will be able to do .S but not .s which will need
support for MASM assembler.