Add default_library=static-for-dll
On Windows when building a static library symbols should not
have dllexport attribute. However, if the static library is ultimately
going to be linked into a DLL, dllexport is needed. Typically projects
do that with a cflag like this:

```
if get_option('default_library') == 'static'
  gst_c_args += ['-DGST_STATIC_COMPILATION']
endif
```

When using `-Ddefault_library=static-for-dll` Meson will still build a
static library, but the snippet above will not define
GST_STATIC_COMPILATION which will cause symbols to be dllexported.