mtest: Display test name in format expected by "meson test" argument
This makes easier to run a single test after running them all. For
example the test "test_accsadubl" in the suite "default" of project
"orc" used to be displayed "orc:default / test_accsadubl". To run that
test specifically the command line is `meson test -C builddir
orc:test_accsadubl --suite default`. To make this more consistent
the test is now displayed as "orc:test_accsadubl (default)".

Fix tests having no suite in the case `suite: []` is passed. Empty list
is now converted to `['']` which was the default value of that kwarg.